requirements.txt
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
et-xmlfile==1.0.1
fire==0.3.1
idna==2.10
jdcal==1.4.1
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.19.1
openpyxl==3.0.4
pandas==1.1.0
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
termcolor==1.1.0
urllib3==1.25.10
